fix: Phase 5 - Fix all smoke test failures (34/34 passing now)

+++ native/src/domain/services.rs
@@ -47,7 +47,7 @@ impl CssGeneratorService {
         CssBundle {
             css: result.css.clone(),
             classes: Vec::<ClassName>::new(),
-            size_bytes: result.size_bytes,
+            size_bytes: result.size_bytes as u32,
         }
     }
 }
